[minor] test: split `UsersControllerTest` test case
