+++ plugins/BEdita/API/tests/TestCase/Controller/UsersControllerTest.php
@@ -230,86 +230,4 @@ class UsersControllerTest extends IntegrationTestCase
         $this->assertArrayHasKey('title', $result['error']);
         $this->assertNotEmpty($result['error']['title']);
     }
-
-    /**
-     * Data provider for `testContentType` test case.
-     *
-     * @return array
-     */
-    public function contentTypeProvider()
-    {
-        return [
-            'json' => [
-                200,
-                'application/json',
-                'application/json',
-            ],
-            'jsonApi' => [
-                200,
-                'application/vnd.api+json',
-                'application/vnd.api+json',
-            ],
-            'jsonApiWrongMediaType' => [
-                415,
-                'application/vnd.api+json',
-                'application/vnd.api+json; m=test',
-            ],
-            'htmlNotAllowed' => [
-                406,
-                null,
-                'text/html,application/xhtml+xml',
-                [
-                    'debug' => 0,
-                    'Accept.html' => 0,
-                ],
-            ],
-            'htmlDebugMode' => [
-                200,
-                'text/html',
-                'text/html,application/xhtml+xml',
-                [
-                    'debug' => 1,
-                    'Accept.html' => 0,
-                ],
-            ],
-            'htmlAccepted' => [
-                200,
-                'text/html',
-                'text/html,application/xhtml+xml',
-                [
-                    'debug' => 0,
-                    'Accept.html' => 1,
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * Test content type negotiation rules.
-     *
-     * @param int $expectedCode Expected response code.
-     * @param string|null $expectedContentType Expected content type.
-     * @param string $accept Request's "Accept" header.
-     * @param array|null $config Configuration to be written.
-     * @return void
-     *
-     * @dataProvider contentTypeProvider
-     * @covers \BEdita\API\Controller\AppController
-     * @covers \BEdita\API\Controller\Component\JsonApiComponent
-     */
-    public function testContentType($expectedCode, $expectedContentType, $accept, array $config = null)
-    {
-        Configure::write($config);
-
-        $this->configRequest([
-            'headers' => ['Accept' => $accept],
-        ]);
-
-        $this->get('/users');
-
-        $this->assertResponseCode($expectedCode);
-        if ($expectedContentType) {
-            $this->assertContentType($expectedContentType);
-        }
-    }
 }
